+++ src/main/java/com/codeborne/selenide/commands/SelectOptionByTextOrIndex.java
@@ -3,30 +3,43 @@ package com.codeborne.selenide.commands;
 import com.codeborne.selenide.Command;
 import com.codeborne.selenide.SelenideElement;
 import com.codeborne.selenide.ex.ElementNotFound;
+import com.codeborne.selenide.ex.InvalidStateException;
+import com.codeborne.selenide.impl.Arguments;
+import com.codeborne.selenide.impl.JavaScript;
 import com.codeborne.selenide.impl.WebElementSource;
-import org.openqa.selenium.NoSuchElementException;
-import org.openqa.selenium.support.ui.Select;
 
 import javax.annotation.Nullable;
 import javax.annotation.ParametersAreNonnullByDefault;
 import java.util.Arrays;
+import java.util.List;
+import java.util.Map;
 
 import static com.codeborne.selenide.Condition.exist;
+import static com.codeborne.selenide.commands.Util.arrayToString;
+import static com.codeborne.selenide.commands.Util.cast;
+import static com.codeborne.selenide.commands.Util.merge;
 
 @ParametersAreNonnullByDefault
 public class SelectOptionByTextOrIndex implements Command<Void> {
+  private static final JavaScript selectOptionByIndex = new JavaScript("select-options-by-index.js");
+  private static final JavaScript selectOptionByText = new JavaScript("select-options-by-text.js");
+
   @Override
   @Nullable
   public Void execute(SelenideElement proxy, WebElementSource selectField, @Nullable Object[] args) {
+    Arguments arguments = new Arguments(args);
     if (args == null || args.length == 0) {
       throw new IllegalArgumentException("Missing arguments");
     }
-    else if (args[0] instanceof String[]) {
-      selectOptionsByTexts(selectField, (String[]) args[0]);
+    else if (args[0] instanceof String) {
+      List<String> texts = merge(arguments.nth(0), arguments.nth(1));
+      selectOptionsByTexts(selectField, texts);
       return null;
     }
-    else if (args[0] instanceof int[]) {
-      selectOptionsByIndexes(selectField, (int[]) args[0]);
+    else if (args[0] instanceof Integer) {
+      Integer index = arguments.nth(0);
+      int[] otherIndexes = arguments.nth(1);
+      selectOptionsByIndexes(selectField, merge(index, otherIndexes));
       return null;
     }
     else {
@@ -34,27 +47,37 @@ public class SelectOptionByTextOrIndex implements Command<Void> {
     }
   }
 
-  private void selectOptionsByTexts(WebElementSource selectField, String[] texts) {
-    Select select = new Select(selectField.getWebElement());
-    for (String text : texts) {
-      try {
-        select.selectByVisibleText(text);
-      }
-      catch (NoSuchElementException e) {
-        throw new ElementNotFound(selectField.getAlias(), selectField.getSearchCriteria() + "/option[text:" + text + ']', exist, e);
-      }
+  private void selectOptionsByTexts(WebElementSource selectField, List<String> texts) {
+    Map<String, String> error = selectOptionByText.execute(selectField.driver(), selectField.getWebElement(), texts);
+    if (error.containsKey("disabledSelect")) {
+      throw new InvalidStateException(selectField.description(), "Cannot select option in a disabled select");
+    }
+    if (error.containsKey("disabledOptions")) {
+      List<String> optionsTexts = cast(error.get("disabledOptions"));
+      String elementDescription = String.format("%s/option[text:%s]", selectField.description(), arrayToString(optionsTexts));
+      throw new InvalidStateException(elementDescription, "Cannot select a disabled option");
+    }
+    if (error.containsKey("optionsNotFound")) {
+      List<String> optionsTexts = cast(error.get("optionsNotFound"));
+      String elementDescription = String.format("%s/option[text:%s]", selectField.getSearchCriteria(), arrayToString(optionsTexts));
+      throw new ElementNotFound(selectField.getAlias(), elementDescription, exist);
     }
   }
 
-  private void selectOptionsByIndexes(WebElementSource selectField, int[] indexes) {
-    Select select = new Select(selectField.getWebElement());
-    for (int index : indexes) {
-      try {
-        select.selectByIndex(index);
-      }
-      catch (NoSuchElementException e) {
-        throw new ElementNotFound(selectField.getAlias(), selectField.getSearchCriteria() + "/option[index:" + index + ']', exist, e);
-      }
+  private void selectOptionsByIndexes(WebElementSource selectField, List<Integer> indexes) {
+    Map<String, Object> error = selectOptionByIndex.execute(selectField.driver(), selectField.getWebElement(), indexes);
+    if (error.containsKey("disabledSelect")) {
+      throw new InvalidStateException(selectField.description(), "Cannot select option in a disabled select");
+    }
+    if (error.containsKey("disabledOptions")) {
+      List<Integer> index = cast(error.get("disabledOptions"));
+      String elementDescription = String.format("%s/option[index:%s]", selectField.description(), arrayToString(index));
+      throw new InvalidStateException(elementDescription, "Cannot select a disabled option");
+    }
+    if (error.containsKey("optionsNotFound")) {
+      List<Integer> index = cast(error.get("optionsNotFound"));
+      String elementDescription = String.format("%s/option[index:%s]", selectField.getSearchCriteria(), arrayToString(index));
+      throw new ElementNotFound(selectField.getAlias(), elementDescription, exist);
     }
   }
 }