+++ src/main/java/com/codeborne/selenide/commands/SelectOptionContainingText.java
@@ -2,48 +2,45 @@ package com.codeborne.selenide.commands;
 
 import com.codeborne.selenide.Command;
 import com.codeborne.selenide.SelenideElement;
+import com.codeborne.selenide.ex.ElementNotFound;
+import com.codeborne.selenide.ex.InvalidStateException;
+import com.codeborne.selenide.impl.Arguments;
+import com.codeborne.selenide.impl.JavaScript;
 import com.codeborne.selenide.impl.WebElementSource;
-import org.openqa.selenium.By;
-import org.openqa.selenium.NoSuchElementException;
-import org.openqa.selenium.WebElement;
-import org.openqa.selenium.support.ui.Quotes;
-import org.openqa.selenium.support.ui.Select;
 
 import javax.annotation.Nullable;
 import javax.annotation.ParametersAreNonnullByDefault;
 import java.util.List;
+import java.util.Map;
 
-import static com.codeborne.selenide.commands.Util.firstOf;
+import static com.codeborne.selenide.Condition.exist;
+import static com.codeborne.selenide.commands.Util.arrayToString;
+import static com.codeborne.selenide.commands.Util.cast;
+import static com.codeborne.selenide.commands.Util.merge;
 
 @ParametersAreNonnullByDefault
 public class SelectOptionContainingText implements Command<Void> {
+  private static final JavaScript selectOptionByPartialText = new JavaScript("select-options-by-partial-text.js");
+
   @Override
   @Nullable
   public Void execute(SelenideElement proxy, WebElementSource selectField, @Nullable Object[] args) {
-    String text = firstOf(args);
-    WebElement element = selectField.getWebElement();
-    Select select = new Select(element);
-
-    List<WebElement> options = element.findElements(By.xpath(
-        ".//option[contains(normalize-space(.), " + Quotes.escape(text) + ")]"));
-
-    if (options.isEmpty()) {
-      throw new NoSuchElementException("Cannot locate option containing text: " + text);
+    Arguments arguments = new Arguments(args);
+    List<String> texts = merge(arguments.nth(0), arguments.nth(1));
+    Map<String, String> error = selectOptionByPartialText.execute(selectField.driver(), selectField.getWebElement(), texts);
+    if (error.containsKey("disabledSelect")) {
+      throw new InvalidStateException(selectField.description(), "Cannot select option in a disabled select");
     }
-
-    for (WebElement option : options) {
-      setSelected(option);
-      if (!select.isMultiple()) {
-        break;
-      }
+    if (error.containsKey("disabledOptions")) {
+      List<String> text = cast(error.get("disabledOptions"));
+      String elementDescription = String.format("%s/option[text containing:%s]", selectField.description(), arrayToString(text));
+      throw new InvalidStateException(elementDescription, "Cannot select a disabled option");
     }
-
+    if (error.containsKey("optionsNotFound")) {
+      List<String> text = cast(error.get("optionsNotFound"));
+      String elementDescription = String.format("%s/option[text containing:%s]", selectField.getSearchCriteria(), arrayToString(text));
+      throw new ElementNotFound(selectField.getAlias(), elementDescription, exist);
+    }
     return null;
-  }
-
-  private void setSelected(WebElement option) {
-    if (!option.isSelected()) {
-      option.click();
-    }
   }
 }