+++ src/main/java/com/codeborne/selenide/commands/Util.java
@@ -9,16 +9,20 @@ import javax.annotation.ParametersAreNonnullByDefault;
 import java.time.Duration;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Objects;
+import java.util.stream.IntStream;
 
 import static java.util.Arrays.asList;
 import static java.util.Collections.emptyList;
+import static java.util.Collections.unmodifiableList;
+import static java.util.stream.Collectors.joining;
 
 @ParametersAreNonnullByDefault
-class Util {
+public class Util {
   @SuppressWarnings("unchecked")
   @CheckReturnValue
   @Nonnull
-  static <T> T firstOf(@Nullable Object[] args) {
+  public static <T> T firstOf(@Nullable Object[] args) {
     if (args == null || args.length == 0) {
       throw new IllegalArgumentException("Missing arguments");
     }
@@ -27,7 +31,7 @@ class Util {
 
   @CheckReturnValue
   @Nonnull
-  static List<Condition> argsToConditions(@Nullable Object[] args) {
+  public static List<Condition> argsToConditions(@Nullable Object[] args) {
     if (args == null) return emptyList();
 
     List<Condition> conditions = new ArrayList<>(args.length);
@@ -40,5 +44,33 @@ class Util {
         throw new IllegalArgumentException("Unknown parameter: " + arg);
     }
     return conditions;
+  }
+
+  @CheckReturnValue
+  @Nonnull
+  @SafeVarargs
+  public static <T> List<T> merge(T first, T... others) {
+    List<T> result = new ArrayList<>(1 + others.length);
+    result.add(first);
+    result.addAll(asList(others));
+    return unmodifiableList(result);
+  }
+
+  @CheckReturnValue
+  @Nonnull
+  public static List<Integer> merge(int first, int[] others) {
+    List<Integer> result = new ArrayList<>(1 + others.length);
+    result.add(first);
+    IntStream.of(others).forEach(i -> result.add(i));
+    return unmodifiableList(result);
+  }
+
+  @SuppressWarnings("unchecked")
+  public static <T> T cast(Object value) {
+    return (T) value;
+  }
+
+  public static <T> String arrayToString(List<T> values) {
+    return values.stream().map(Objects::toString).collect(joining(","));
   }
 }