+++ src/main/java/com/codeborne/selenide/impl/DownloadFileToFolder.java
@@ -18,7 +18,7 @@ import javax.annotation.ParametersAreNonnullByDefault;
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.util.List;
-import java.util.function.Predicate;
+import java.util.Map;
 
 import static com.codeborne.selenide.impl.FileHelper.moveFile;
 import static java.lang.System.currentTimeMillis;
@@ -30,6 +30,8 @@ import static java.util.stream.Collectors.toList;
 @ParametersAreNonnullByDefault
 public class DownloadFileToFolder {
   private static final Logger log = LoggerFactory.getLogger(DownloadFileToFolder.class);
+  private static final String CHROME_TEMPORARY_FILE = "crdownload";
+  private static final String FIREFOX_TEMPORARY_FILE = "part";
 
   private final Downloader downloader;
   private final Waiter waiter;
@@ -78,8 +80,14 @@ public class DownloadFileToFolder {
 
     action.perform(driver, clickable);
 
-    waitUntilDownloadsCompleted(driver.browser(), timeout, pollingInterval, folder);
-    Downloads newDownloads = waitForNewFiles(timeout, fileFilter, config, folder, downloadStartedAt);
+    waitForNewFiles(timeout, fileFilter, config, folder, downloadStartedAt);
+    waitUntilDownloadsCompleted(driver.browser(), folder, timeout, pollingInterval);
+    Downloads newDownloads = new Downloads(newFiles(folder, downloadStartedAt));
+
+    if (log.isInfoEnabled()) {
+      log.info("Downloaded {}", newDownloads.filesAsString());
+    }
+
     File downloadedFile = newDownloads.firstDownloadedFile(anyClickableElement.toString(), timeout, fileFilter);
 
     if (log.isDebugEnabled()) {
@@ -88,40 +96,58 @@ public class DownloadFileToFolder {
     return archiveFile(config, downloadedFile);
   }
 
-  private void waitUntilDownloadsCompleted(Browser browser, long timeout, long pollingInterval, DownloadsFolder folder) {
-    pause(pollingInterval);
+  private void waitUntilDownloadsCompleted(Browser browser, DownloadsFolder folder,
+                                           long timeout, long pollingInterval) {
     if (browser.isChrome() || browser.isEdge()) {
-      waitUntilFileDisappears(folder, "crdownload", timeout, pollingInterval);
+      waitUntilFileDisappears(folder, CHROME_TEMPORARY_FILE, timeout, pollingInterval);
     }
     else if (browser.isFirefox()) {
-      waitUntilFileDisappears(folder, "part", timeout, pollingInterval);
+      waitUntilFileDisappears(folder, FIREFOX_TEMPORARY_FILE, timeout, pollingInterval);
+    }
+    else {
+      waitWhileFilesAreBeingModified(folder, timeout, pollingInterval);
     }
   }
 
   private void waitUntilFileDisappears(DownloadsFolder folder, String extension, long timeout, long pollingInterval) {
-    for (long start = currentTimeMillis(); currentTimeMillis() - start < timeout && folder.hasFiles(extension); ) {
-      log.info("Found {} files in {}, waiting...", extension, folder);
-      pause(pollingInterval);
-    }
+    waiter.wait(timeout, pollingInterval, () -> {
+      log.debug("Found {} files in {}, waiting for {} ms...", extension, folder, pollingInterval);
+      return !folder.hasFiles(extension);
+    });
 
     if (folder.hasFiles(extension)) {
-      log.warn("Folder {} still contains files {}", folder, extension);
+      log.warn("Folder {} still contains files {} after {} ms.", folder, extension, timeout);
     }
   }
 
-  @Nonnull
-  private Downloads waitForNewFiles(long timeout, FileFilter fileFilter, Config config,
-                                    DownloadsFolder folder, long clickMoment) {
-    HasDownloads hasDownloads = new HasDownloads(fileFilter, clickMoment);
-    waiter.wait(folder, hasDownloads, timeout, config.pollingInterval());
+  private void waitWhileFilesAreBeingModified(DownloadsFolder folder, long timeout, long pollingInterval) {
+    Map<String, Long> times = folder.modificationTimes();
+    long lastModifiedAt = currentTimeMillis();
 
-    if (log.isInfoEnabled()) {
-      log.info(hasDownloads.downloads.filesAsString());
+    for (long start = currentTimeMillis(); currentTimeMillis() - start < timeout; pause(pollingInterval)) {
+      var newTimes = folder.modificationTimes();
+      if (!newTimes.equals(times)) {
+        log.debug("Files has been modified - old: {}, new: {}", times, newTimes);
+        lastModifiedAt = currentTimeMillis();
+        times = newTimes;
+      }
+      else {
+        log.debug("Files has not been modified in last {} ms: {}", pollingInterval, times);
+        if (currentTimeMillis() - lastModifiedAt > 1000) {
+          log.debug("Files has not been modified during last {} ms.", currentTimeMillis() - lastModifiedAt);
+          return;
+        }
+      }
     }
-    if (log.isDebugEnabled()) {
-      log.debug("All downloaded files in {}: {}", folder, folder.files().stream().map(f -> f.getName()).collect(joining("\n")));
-    }
-    return hasDownloads.downloads;
+    log.warn("Files are still being modified during last {} ms.", currentTimeMillis() - lastModifiedAt);
+  }
+
+  private void waitForNewFiles(long timeout, FileFilter fileFilter, Config config,
+                                    DownloadsFolder folder, long clickMoment) {
+    waiter.wait(timeout, config.pollingInterval(), () -> {
+      Downloads downloads = new Downloads(newFiles(folder, clickMoment));
+      return !downloads.files(fileFilter).isEmpty();
+    });
   }
 
   private void pause(long milliseconds) {
@@ -142,30 +168,12 @@ public class DownloadFileToFolder {
     return archivedFile;
   }
 
-  @ParametersAreNonnullByDefault
-  private static class HasDownloads implements Predicate<DownloadsFolder> {
-    private final FileFilter fileFilter;
-    private final long downloadStartedAt;
-    Downloads downloads;
-
-    private HasDownloads(FileFilter fileFilter, long downloadStartedAt) {
-      this.fileFilter = fileFilter;
-      this.downloadStartedAt = downloadStartedAt;
-    }
-
-    @Override
-    public boolean test(DownloadsFolder folder) {
-      downloads = new Downloads(newFiles(folder));
-      return !downloads.files(fileFilter).isEmpty();
-    }
-
-    private List<DownloadedFile> newFiles(DownloadsFolder folder) {
-      return folder.files().stream()
-        .filter(File::isFile)
-        .filter(file -> isFileModifiedLaterThan(file, downloadStartedAt))
-        .map(file -> new DownloadedFile(file, emptyMap()))
-        .collect(toList());
-    }
+  private static List<DownloadedFile> newFiles(DownloadsFolder folder, long modifiedAfterTs) {
+    return folder.files().stream()
+      .filter(File::isFile)
+      .filter(file -> isFileModifiedLaterThan(file, modifiedAfterTs))
+      .map(file -> new DownloadedFile(file, emptyMap()))
+      .collect(toList());
   }
 
   /**