+++ src/main/resources/select-options-by-partial-text.js
@@ -1,0 +1,30 @@
+(function (select, texts) {
+  if (select.disabled) {
+    return {disabledSelect: 'Cannot select option in a disabled select'};
+  }
+
+  function optionByText(requestedText) {
+    return Array.from(select.options).find(option => option.text.includes(requestedText))
+  }
+
+  const missingOptionsTexts = texts.filter(text => !optionByText(text));
+  if (missingOptionsTexts.length > 0) {
+    return {optionsNotFound: missingOptionsTexts};
+  }
+
+  const disabledOptionsTexts = texts.filter(text => optionByText(text).disabled);
+  if (disabledOptionsTexts.length > 0) {
+    return {disabledOptions: disabledOptionsTexts};
+  }
+
+  for (let requestedPartialText of texts) {
+    optionByText(requestedPartialText).selected = 'selected';
+  }
+
+  const event = document.createEvent('HTMLEvents');
+  event.initEvent('change', true, true);
+  select.dispatchEvent(event);
+
+  return {};
+})(arguments[0], arguments[1])
+