+++ src/test/java/integration/server/BaseHandler.java
@@ -17,6 +17,8 @@ import java.util.Map;
 import java.util.Objects;
 import java.util.stream.Stream;
 
+import static integration.server.Delayer.writeSlowly;
+import static integration.server.Delayer.pause;
 import static java.lang.Thread.currentThread;
 import static java.nio.charset.StandardCharsets.UTF_8;
 import static java.util.stream.Collectors.joining;
@@ -24,6 +26,7 @@ import static javax.servlet.http.HttpServletResponse.SC_BAD_REQUEST;
 import static javax.servlet.http.HttpServletResponse.SC_METHOD_NOT_ALLOWED;
 
 public abstract class BaseHandler extends HttpServlet {
+  static final String CONTENT_TYPE_PLAIN_TEXT = "text/plain";
   static final String CONTENT_TYPE_HTML_TEXT = "text/html";
   private static final String CONTENT_TYPE_IMAGE_PNG = "image/png";
   private static final Logger log = LoggerFactory.getLogger(BaseHandler.class);
@@ -49,6 +52,10 @@ public abstract class BaseHandler extends HttpServlet {
   }
 
   private void result(HttpServletRequest request, HttpServletResponse response, long start, Result result) throws IOException {
+    if (result.pause > 0) {
+      log.debug("sleep before [{}] for {} ms", request.getPathInfo(), result.pause);
+      pause(result.pause);
+    }
     response.setStatus(result.httpStatus);
     response.setContentLength(result.content.length);
     response.setContentType(result.contentType);
@@ -56,7 +63,12 @@ public abstract class BaseHandler extends HttpServlet {
       response.setHeader(httpHeader.getKey(), httpHeader.getValue());
     }
     try (OutputStream os = response.getOutputStream()) {
-      os.write(result.content);
+      if (result.duration == 0) {
+        os.write(result.content);
+      }
+      else {
+        writeSlowly(os, result.content, result.duration);
+      }
     }
     logRequest(request, result.httpStatus, start);
     if (result.httpStatus >= SC_BAD_REQUEST) {
@@ -82,13 +94,14 @@ public abstract class BaseHandler extends HttpServlet {
   @SuppressWarnings("SuspiciousDateFormat")
   private void logRequest(HttpServletRequest request, int httpStatus, long startTime) {
     String time = new SimpleDateFormat("hh:MM:ss:SSS").format(new Date());
-    log.info("{} {} -> {} {} ms",
+    log.info("{} {} {} -> {} {} ms",
       time,
+      request.getMethod(),
       Stream.of(request.getRequestURL(), request.getQueryString())
         .filter(Objects::nonNull)
         .collect(joining("?")),
       httpStatus,
-      (System.nanoTime() - startTime) / 1000000
+      (System.nanoTime() - startTime) / 1_000_000
     );
   }
 }