+++ src/test/java/integration/server/FileDownloadHandler.java
@@ -9,7 +9,9 @@ import java.util.Arrays;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Set;
+import java.util.stream.Collectors;
 
+import static java.lang.Long.parseLong;
 import static javax.servlet.http.HttpServletResponse.SC_NOT_FOUND;
 import static javax.servlet.http.HttpServletResponse.SC_OK;
 import static javax.servlet.http.HttpServletResponse.SC_UNAUTHORIZED;
@@ -34,16 +36,6 @@ class FileDownloadHandler extends BaseHandler {
       return new Result(SC_NOT_FOUND, CONTENT_TYPE_HTML_TEXT, "NOT_FOUND: " + fileName);
     }
 
-    if (request.getParameter("pause") != null) {
-      try {
-        Thread.sleep(Long.parseLong(request.getParameter("pause")));
-      }
-      catch (InterruptedException e) {
-        Thread.currentThread().interrupt();
-        throw new RuntimeException(e);
-      }
-    }
-
     String contentType = getContentType(fileName);
     if ("файл-с-запрещёнными-символами.txt".equals(fileName)) {
       fileName = "имя с #pound,%percent,&ampersand,{left,}right,\\backslash," +
@@ -53,17 +45,24 @@ class FileDownloadHandler extends BaseHandler {
     }
     Map<String, String> map = new HashMap<>();
     map.put("content-disposition", "attachment; filename=" + URLEncoder.encode(fileName, "UTF-8"));
-    return new Result(SC_OK, contentType, fileContent, map);
+    return new Result(SC_OK, contentType, fileContent, map, longParam(request, "pause"), longParam(request, "duration"));
+  }
+
+  private long longParam(HttpServletRequest request, String name) {
+    String param = request.getParameter(name);
+    return param == null ? 0 : parseLong(param);
   }
 
   private String getSessionId(HttpServletRequest request) {
-    if (request.getCookies() != null) {
-      for (Cookie cookie : request.getCookies()) {
-        if ("session_id".equals(cookie.getName())) {
-          return cookie.getValue();
-        }
+    if (request.getCookies() == null) {
+      return "Request has no cookies";
+    }
+    for (Cookie cookie : request.getCookies()) {
+      if ("session_id".equals(cookie.getName())) {
+        return cookie.getValue();
       }
     }
-    throw new IllegalArgumentException("No cookie 'session_id' found: " + Arrays.toString(request.getCookies()));
+
+    return "No cookie 'session_id' found: " + Arrays.stream(request.getCookies()).map(Cookie::getName).collect(Collectors.toList());
   }
 }