+++ statics/src/test/java/integration/UnopenedBrowserTest.java
@@ -19,7 +19,7 @@ import static com.codeborne.selenide.Selenide.closeWebDriver;
 import static com.codeborne.selenide.Selenide.open;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.junit.jupiter.api.Assumptions.assumeTrue;
+import static org.assertj.core.api.Assumptions.assumeThat;
 
 final class UnopenedBrowserTest extends IntegrationTest {
   @BeforeEach
@@ -66,13 +66,13 @@ final class UnopenedBrowserTest extends IntegrationTest {
     ElementsCollection headers = $$("h2");
     openFile("page_with_selects_without_jquery.html");
     headers
-      .shouldHave(size(3))
-      .shouldHave(texts("Dropdown list", "Options with 'apostrophes' and \"quotes\"", "Radio buttons"));
+      .shouldHave(size(4))
+      .shouldHave(texts("Dropdown list", "Options with 'apostrophes' and \"quotes\"", "Radio buttons", "Disabled select"));
   }
 
   @Test
   void canDeclareSelenideElements_beforeOpeningBrowsers_evenIfUsingCustomWebdriver() {
-    assumeTrue(browser().isChrome());
+    assumeThat(browser().isChrome()).isTrue();
     useProxy(false);
 
     SelenideElement header = $("h1");